lesson11

+++ app/src/main/java/com/example/yournote/SocialNetworkFragment.java
@@ -1,16 +1,14 @@
 package com.example.yournote;
 
-import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.DividerItemDecoration;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.os.Parcelable;
 import android.view.ContextMenu;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -18,43 +16,24 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 import android.widget.Toast;
-import com.example.yournote.R;
 
 
 public class SocialNetworkFragment extends Fragment {
-    public static final String ARG_INDEX = "index";
     SocialNetworkAdapter adapter;
     CardSource data;
     RecyclerView recyclerView;
-    CardData newCard;
+    private Navigation navigation;
+    private final Publisher publisher = new Publisher();
 
     public static SocialNetworkFragment newInstance() {
         return new SocialNetworkFragment();
     }
-//todo
-//    public static SocialNetworkFragment newInst(SocialNetworkFragment f) {
-//        Bundle args = new Bundle();
-//        args.putParcelable(ARG_INDEX, new CardData(null, null, null));
-//        f.setArguments(args);
-//        return f;
-//    }
-//todo
-//    @Override
-//    public void onCreate(Bundle savedInstanceState) {
-//        super.onCreate(savedInstanceState);
-//        if (getArguments() != null) {
-//            newCard = getArguments().getParcelable(ARG_INDEX);
-//            for (int i = 0; i < data.size(); i++) {
-//                if (newCard.getNoteName()==data.getData(i).getNoteName()){
-//                    data.getData(i).setNote(newCard.getNote());
-//                    data.getData(i).setDates(newCard.getDates());
-//                    adapter.notifyItemChanged(i);
-//                }
-//            }
-//        }
-//    }
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        data = new CardSourceImpl(getResources()).init();
+    }
 
 
     @Override
@@ -63,39 +42,35 @@ public class SocialNetworkFragment extends Fragment {
         View view = inflater.inflate(R.layout.fragment_social_network, container, false);
         recyclerView = view.findViewById(R.id.recycler_view_lines);
         data = new CardSourceImpl (getResources()).init();
-        initView(view);
+        initRecyclerView();
         setHasOptionsMenu(true);
         return view;
     }
 
     @Override
-    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+    public void onCreateOptionsMenu(@NonNull Menu menu, MenuInflater inflater) {
                 inflater.inflate(R.menu.main_menu, menu);
     }
 
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_add:
-                data.addCardData(new CardData("Заголовок " + data.size(),
-                        "Описание " + data.size(), "сделать завтра"));
-                adapter.notifyItemInserted(data.size() - 1);
-                recyclerView.scrollToPosition(data.size() - 1);
-                return true;
-            case R.id.action_clear:
-                data.clearCardData();
-                adapter.notifyDataSetChanged();
-                return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-    private void initView(View view) {
-        recyclerView = view.findViewById(R.id.recycler_view_lines);
-        data = new CardSourceImpl(getResources()).init();
-        initRecyclerView();
-    }
+//    @Override
+//    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+//        switch (item.getItemId()){
+//            case R.id.action_add:
+//                navigation.addFragment(TextFragment.newInstance());
+//                publisher.subscribe(cardData -> {
+//                    data.addCardData(cardData);
+//                    adapter.notifyItemInserted(data.size() - 1);
+//                });
+//                return true;
+//            case R.id.action_clear:
+//                data.clearCardData();
+//                adapter.notifyDataSetChanged();
+//                return true;
+//        }
+//        return super.onOptionsItemSelected(item);
+//    }
 
-    private void initRecyclerView(){
+    private void initRecyclerView() {
         recyclerView.setHasFixedSize(true);
 
         LinearLayoutManager layoutManager = new LinearLayoutManager(getContext());
@@ -114,32 +89,32 @@ public class SocialNetworkFragment extends Fragment {
 
     @Override
     public boolean onContextItemSelected(@NonNull MenuItem item) {
-        int position = adapter.getMenuPosition();
-        CardData card = data.getData(position);
+        final int position = adapter.getMenuPosition();
         switch(item.getItemId()) {
             case R.id.action_update:
-                showTextFra(card);
-//                data.updateCardData(position,
-//                        new CardData(newCard.getNoteName(),
-//                                newCard.getNote(),
-//                                newCard.getDates()));
-//                adapter.notifyItemChanged(position);
-
+                navigation.addFragment(TextFragment.newInstance(data.getCardData(position)), true);
+                publisher.subscribe(cardData -> {
+                    data.updateCardData(position, cardData);
+                    adapter.notifyItemChanged(position);
+                });
                 return true;
             case R.id.action_delete:
-                data.deleteCardData(position);
-                adapter.notifyItemRemoved(position);
-                return true;
+                    AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
+                    builder.setTitle(R.string.exclamation)
+                            .setMessage(R.string.question)
+                            .setCancelable(false)
+                            .setNegativeButton(R.string.no,
+                                    (dialog, id) -> Toast.makeText(getContext(), "Заметка не удалена",
+                                            Toast.LENGTH_SHORT).show())
+                            .setPositiveButton(R.string.yes,
+                                    (dialog, id) -> {
+                                        data.deleteCardData(position);
+                                        adapter.notifyItemRemoved(position);
+                                    });
+
+                    AlertDialog alert = builder.create();
+                    alert.show();
         }
         return super.onContextItemSelected(item);
     }
-
-    private void showTextFra(CardData card) {
-        Intent intent = new Intent();
-        intent.setClass(getActivity(), TextActivity.class);
-        //intent.putParcelableArrayListExtra(TextFragment.ARG_INDEX, card);
-        intent.putExtra(TextFragment.ARG_INDEX, (Parcelable) card);
-        startActivity(intent);
-    }
-
 }
